CAMS-97: Move calculateDifferenceInMonths to date-helper
