+++ backend/functions/lib/adapters/gateways/cases.local.gateway.test.ts
@@ -1,5 +1,6 @@
 import { CasesLocalGateway } from './cases.local.gateway';
 import { GatewayHelper } from './gateway-helper';
+import { calculateDifferenceInMonths } from '../utils/date-helper';
 
 const context = require('azure-function-context-mock');
 let gatewayHelper: GatewayHelper;
@@ -12,7 +13,7 @@ describe('CasesLocalGateway tests', () => {
   test('should return data when passed starting month', async () => {
     const casesLocalGateway = new CasesLocalGateway();
     const startDate = new Date(2015, 1, 1);
-    const startingMonth = 0 - gatewayHelper.calculateDifferenceInMonths(new Date(), startDate);
+    const startingMonth = 0 - calculateDifferenceInMonths(new Date(), startDate);
     const cases = await casesLocalGateway.getChapter15Cases(context, { startingMonth });
     expect(cases).toBeTruthy();
     expect(cases).not.toEqual([]);
