+++ backend/functions/lib/adapters/gateways/gateway-helper.ts
@@ -25,30 +25,4 @@ export class GatewayHelper {
       throw Error(err);
     }
   }
-
-  calculateDifferenceInMonths(left: Date, right: Date): number {
-    let earlier, later: Date;
-    if (left.getFullYear() < right.getFullYear()) {
-      earlier = left;
-      later = right;
-    } else if (left.getFullYear() > right.getFullYear()) {
-      earlier = right;
-      later = left;
-    } else if (left.getMonth() < right.getMonth()) {
-      earlier = left;
-      later = right;
-    } else if (left.getMonth() > right.getMonth()) {
-      earlier = right;
-      later = left;
-    } else {
-      return 0;
-    }
-    const years = Math.abs(earlier.getFullYear() - later.getFullYear());
-    const incompleteYear = later.getMonth() < earlier.getMonth();
-    const months = later.getMonth() - earlier.getMonth();
-    const realMonths = months < 0 ? 12 + months : months;
-    const incompleteMonth = later.getDate() < earlier.getDate();
-    const monthsDiff = incompleteYear ? years * 12 + realMonths - 12 : years * 12 + realMonths;
-    return incompleteMonth ? monthsDiff - 1 : monthsDiff;
-  }
 }
